Fix some code indentation and utils renamed
