+++ react-native.config.js
@@ -1,4 +1,4 @@
 // Import app fonts
 module.exports = {
-  assets: ["./src/fonts/"],
-};
+    assets: ['./src/fonts/'],
+}
