+++ src/api/contents.js
@@ -1,4 +1,4 @@
-import { API_URL } from 'react-native-dotenv';
+import { API_URL } from 'react-native-dotenv'
 
 export const getContents = async (token) => {
     const response = await fetch(`${API_URL}/contents`, {
@@ -6,7 +6,7 @@ export const getContents = async (token) => {
             Accept: 'application/vnd.api+json',
             Authorization: token,
         },
-    });
+    })
     return {
         status: response.status,
         body: await response.json(),
