+++ src/navigation/AppStack.js
@@ -24,7 +24,7 @@ import Tutorial from '../screens/principal/Tutorial'
 import Vigilancia from '../screens/principal/Vigilancia'
 
 import translate from '../../locales/i18n'
-import { scale } from '../utils/scallingUtils'
+import { scale } from '../utils/scalling'
 
 const Stack = createStackNavigator()
 Feather.loadFont()
