+++ src/screens/inicio/Login/index.js
@@ -19,7 +19,7 @@ import { Logo, PageTitle, LabelVisible } from './styles'
 
 import translate from '../../../../locales/i18n'
 import { GDSLogoES, GDSLogoBR } from '../../../img/imageConst'
-import { scale } from '../../../utils/scallingUtils'
+import { scale } from '../../../utils/scalling'
 import { useUser } from '../../../hooks/user'
 import { authUser } from '../../../api/user'
