+++ src/screens/principal/Ajuda/index.js
@@ -3,7 +3,7 @@ import React from 'react'
 import { Help, Box, Button, IconWrapper, InfoWrapper, Title } from './styles'
 
 import translate from '../../../../locales/i18n'
-import { scale } from '../../../utils/scallingUtils'
+import { scale } from '../../../utils/scalling'
 import { TermsIcon, TutorialIcon, InfoIcon } from '../../../img/imageConst'
 
 const Ajuda = ({ navigation }) => {
