+++ src/screens/principal/EditarPerfil/index.js
@@ -35,18 +35,18 @@ import { Delete } from './styles'
 import InstitutionSelector from '../../../components/Groups/InstitutionSelector'
 import LoadingModal from '../../../components/Groups/LoadingModal'
 import translate from '../../../../locales/i18n'
-import { scale } from '../../../utils/scallingUtils'
+import { scale } from '../../../utils/scalling'
 import {
     genderChoices,
     countryChoices,
     raceChoices,
     householdChoices,
-} from '../../../utils/selectorUtils'
+} from '../../../utils/selector'
 import {
     handleAvatar,
     getInitials,
     validatePerson,
-} from '../../../utils/constUtils'
+} from '../../../utils/consts'
 import { stateOptions, getCity } from '../../../utils/brasil'
 import { useUser } from '../../../hooks/user'
 import { updateUser } from '../../../api/user'
