+++ src/screens/principal/EditarPerfil/styles.js
@@ -2,7 +2,7 @@ import styled from 'styled-components'
 
 import { TouchableOpacity } from 'react-native'
 
-import { scale } from '../../../utils/scallingUtils'
+import { scale } from '../../../utils/scalling'
 
 export const Delete = styled(TouchableOpacity).attrs({
     activeOpacity: 0.5,
