+++ src/screens/principal/Noticias/styles.js
@@ -2,7 +2,7 @@ import styled from 'styled-components/native'
 
 import ShadowView from 'react-native-simple-shadow-view'
 
-import { scale, percentage } from '../../../utils/scallingUtils'
+import { scale, percentage } from '../../../utils/scalling'
 
 export const Container = styled.View`
     background-color: #348eac;
