+++ src/screens/principal/NovoPerfil/index.js
@@ -32,14 +32,14 @@ import {
 import InstitutionSelector from '../../../components/Groups/InstitutionSelector'
 import LoadingModal from '../../../components/Groups/LoadingModal'
 import translate from '../../../../locales/i18n'
-import { scale } from '../../../utils/scallingUtils'
+import { scale } from '../../../utils/scalling'
 import {
     genderChoices,
     countryChoices,
     raceChoices,
     householdChoices,
-} from '../../../utils/selectorUtils'
-import { validatePerson } from '../../../utils/constUtils'
+} from '../../../utils/selector'
+import { validatePerson } from '../../../utils/consts'
 import { useUser } from '../../../hooks/user'
 import { createHousehold } from '../../../api/households'
