+++ src/screens/principal/Perfis/index.js
@@ -24,8 +24,8 @@ import {
 } from './styles'
 
 import translate from '../../../../locales/i18n'
-import { scale } from '../../../utils/scallingUtils'
-import { handleAvatar, getInitials } from '../../../utils/constUtils'
+import { scale } from '../../../utils/scalling'
+import { handleAvatar, getInitials } from '../../../utils/consts'
 import { useUser } from '../../../hooks/user'
 import { getUser } from '../../../api/user'
