+++ src/screens/principal/Rumor/index.js
@@ -23,7 +23,7 @@ import { CoolAlert } from '../../../components/CoolAlert'
 import { ExitMap, ConfirmMap, MapFormMarker, MapFormText } from './styles'
 
 import translate from '../../../../locales/i18n'
-import { scale } from '../../../utils/scallingUtils'
+import { scale } from '../../../utils/scalling'
 import { useUser } from '../../../hooks/user'
 import { createRumor } from '../../../api/rumors'
