+++ src/screens/principal/Rumor/styles.js
@@ -2,7 +2,7 @@ import styled from 'styled-components'
 
 import { TouchableOpacity } from 'react-native'
 
-import { scale } from '../../../utils/scallingUtils'
+import { scale } from '../../../utils/scalling'
 
 export const ExitMap = styled(TouchableOpacity).attrs({
     activeOpacity: 0.2,
