+++ src/screens/principal/Sobre/index.js
@@ -6,8 +6,8 @@ import { ScrollViewStyled, Title, ImageContainer, BodyText } from './styles'
 
 import translate from '../../../../locales/i18n'
 import { UnBLogo, ProEpiLogo } from '../../../img/imageConst'
-import { Redirect } from '../../../utils/constUtils'
-import { scale } from '../../../utils/scallingUtils'
+import { redirectAlert } from '../../../utils/consts'
+import { scale } from '../../../utils/scalling'
 
 const Sobre = () => {
     return (
@@ -19,7 +19,7 @@ const Sobre = () => {
                 <ImageContainer>
                     <TouchableOpacity
                         onPress={() =>
-                            Redirect(
+                            redirectAlert(
                                 translate('about.tituloBtnProEpi'),
                                 translate('about.mensagemBtnProEpi'),
                                 translate('about.linkBtnProEPi')
@@ -34,7 +34,7 @@ const Sobre = () => {
 
                     <TouchableOpacity
                         onPress={() =>
-                            Redirect(
+                            redirectAlert(
                                 translate('about.tituloBtnUnb'),
                                 translate('about.mensagemBtnUnb'),
                                 translate('about.linkBtnUnb')
