+++ src/screens/principal/Vigilancia/index.js
@@ -25,7 +25,7 @@ import { Title, BodyText } from './styles'
 
 import translate from '../../../../locales/i18n'
 import { useUser } from '../../../hooks/user'
-import { scale } from '../../../utils/scallingUtils'
+import { scale } from '../../../utils/scalling'
 import { updateUser } from '../../../api/user'
 
 Feather.loadFont()
